Final commit

+++ Controllers/IotDataController.cs
@@ -17,6 +17,8 @@ namespace DotnetAssignment1.Controllers
             _logger = logger;
         }
 
+        /// <summary>
+        // Retrieves the value associated with the specified key from the IoT data repository.
         [HttpGet("/keys/{key}")]
         public async Task<ActionResult<Response>> GetValueByKey(string key)
         {
@@ -24,25 +26,28 @@ namespace DotnetAssignment1.Controllers
             {
                 if (string.IsNullOrEmpty(key))
                 {
-                    return BadRequest(new Response { Message = "Enter valid key" });
+                    return StatusCode(400,new Response { Message = "Enter valid key", StatusCode=400 });
                 }
 
                 var keyValue = await _iotDataRepository.GetValueByKeyAsync(key);
 
                 if (keyValue == null)
                 {
-                    return NotFound(new Response { Message = "No value exists for the given key" });
+                    return StatusCode(404,new Response { Message = "No value exists for the given key",StatusCode = 404 });
                 }
 
-                return Ok(new Response { Message = "Successfully fetched value for the corresponding key", Data = keyValue });
+                return StatusCode(200,new Response { Message = "Successfully fetched value for the corresponding key", StatusCode = 200,Data = keyValue });
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "An error occurred while getting value by key: {Key}", key);
-                return StatusCode(StatusCodes.Status500InternalServerError, new Response { Message = "Internal server error" });
+                return StatusCode(StatusCodes.Status500InternalServerError, new Response { Message = "Internal server error",StatusCode = 500 });
             }
         }
 
+
+        /// <summary>
+        // Adds a new key-value pair to the IoT data repository.
         [HttpPost("/keys")]
         public async Task<ActionResult<Response>> AddKeyValue([FromBody] AddKeyValue request)
         {
@@ -50,7 +55,7 @@ namespace DotnetAssignment1.Controllers
             {
                 if (request == null)
                 {
-                    return BadRequest(new Response { Message = "Invalid data. Please enter correct data", StatusCode = 400 });
+                    return StatusCode(400,new Response { Message = "Invalid data. Please enter correct data", StatusCode = 400 });
                 }
 
                 var response = await _iotDataRepository.AddKeyValueAsync(request);
@@ -64,13 +69,18 @@ namespace DotnetAssignment1.Controllers
             }
         }
 
+        /// <summary>
+        // Update Key with new value  to the IoT data repository.
         [HttpPatch("/keys/{key}/{value}")]
         public async Task<ActionResult<Response>> UpdateValue(string key, string value)
         {
             try
             {
+                if (key == null || value == null)
+                {
+                    return StatusCode(400, new Response { StatusCode = 400, Message = "Invalid data. Please enter correct data" });
+                }
                 var response = await _iotDataRepository.UpdateValueAsync(key, value);
-
                 return StatusCode(response.StatusCode, response);
             }
             catch (Exception ex)
@@ -80,13 +90,22 @@ namespace DotnetAssignment1.Controllers
             }
         }
 
+
+
+
+        /// <summary>
+        // Delete key-value pair to the IoT data repository.
         [HttpDelete("/keys/{key}")]
         public async Task<ActionResult<Response>> DeleteKeyValue(string key)
         {
             try
             {
-                var response = await _iotDataRepository.DeleteKeyValueAsync(key);
 
+                if (key == null )
+                {
+                    return StatusCode(400, new Response { StatusCode = 400, Message = "Invalid data. Please enter correct data" });
+                }
+                var response = await _iotDataRepository.DeleteKeyValueAsync(key);
                 return StatusCode(response.StatusCode, response);
             }
             catch (Exception ex)
